chore: component 폴더로 이동
