+++ src/pages/ChattingPage/components/ChatRoomTitle.tsx
@@ -1,0 +1,35 @@
+import { ChatRoomDetail } from '@type/models/ChatRoom';
+
+import { CHAT_ROOM_TAB_TITLE } from '@consts/chat';
+import { PATH_NAME } from '@consts/pathName';
+
+import { Pointer } from '../ChattingPage.style';
+
+export const ChatRoomTitle = ({
+  type,
+  domainId,
+  roomName,
+  onClick: moveToPage,
+}: Pick<ChatRoomDetail, 'type' | 'domainId' | 'roomName'> & {
+  onClick: (path: string) => void;
+}) => {
+  const titlePagePath = getTitlePagePath(type);
+
+  return (
+    <Pointer onClick={() => moveToPage(titlePagePath(String(domainId)))}>
+      {roomName}
+    </Pointer>
+  );
+};
+
+const getTitlePagePath = (roomType: string) => {
+  if (roomType === CHAT_ROOM_TAB_TITLE.INDIVIDUAL) {
+    return PATH_NAME.GET_PROFILE_PATH;
+  }
+
+  if (roomType === CHAT_ROOM_TAB_TITLE.GUEST) {
+    return PATH_NAME.GET_GAMES_PATH;
+  }
+
+  return PATH_NAME.GET_CREWS_PATH;
+};
